scripts/requirements.txt
Pillow>=10.0
pyzbar>=0.1.9
supabase>=2.0
python-dotenv>=1.0
rich>=13.0
click>=8.0
